APPCLI-9: consolidate configuration management and more work on migration.
