+++ appcli/commands/configure_cli.py
@@ -10,35 +10,26 @@ www.brightsparklabs.com
 """
 
 # standard library
-import json
 import os
-import shutil
-import tarfile
-from datetime import datetime, timezone
-from pathlib import Path
 from typing import Iterable
 
 # vendor libraries
 import click
-from jinja2 import StrictUndefined, Template
 
 # local libraries
-from appcli.configuration_manager import ConfigurationManager
-from appcli.crypto.crypto import create_and_save_key, decrypt_values_in_file
-from appcli.functions import (
-    error_and_exit,
-    get_generated_configuration_metadata_file,
-    print_header,
-    execute_validation_functions,
-)
-from appcli.git_repositories.git_repositories import (
-    ConfigurationGitRepository,
-    GeneratedConfigurationGitRepository,
+from appcli.configuration_manager import (
+    ConfigurationManager,
     confirm_config_dir_exists,
     confirm_config_dir_not_exists,
     confirm_generated_config_dir_exists,
     confirm_generated_config_dir_is_not_dirty,
     confirm_generated_configuration_is_using_current_configuration,
+    confirm_not_on_master_branch,
+)
+from appcli.functions import (
+    error_and_exit,
+    print_header,
+    execute_validation_functions,
 )
 from appcli.logger import logger
 from appcli.models.cli_context import CliContext
@@ -79,22 +70,20 @@ class ConfigureCli:
 
             cli_context: CliContext = ctx.obj
 
+            # Validate environment
             self.__pre_configure_init_validation(cli_context)
-
             self.__check_env_vars_set(cli_context, self.mandatory_env_variables)
 
+            # Run pre-hooks
             hooks = self.cli_configuration.hooks
-
             logger.debug("Running pre-configure init hook")
             hooks.pre_configure_init(ctx)
 
             # Initialise configuration directory
             logger.debug("Initialising configuration directory")
-            self.__initialise_configuration_directory(cli_context)
+            ConfigurationManager(cli_context, self.cli_configuration).init()
 
-            # Create an encryption key
-            create_and_save_key(cli_context.get_key_file())
-
+            # Run post-hooks
             logger.debug("Running post-configure init hook")
             hooks.post_configure_init(ctx)
 
@@ -117,27 +106,21 @@ class ConfigureCli:
         def apply(ctx, message, force):
             cli_context: CliContext = ctx.obj
 
+            # Validate environment
             self.__pre_configure_apply_validation(cli_context, force=force)
 
-            configuration = ConfigurationManager(
-                cli_context.get_app_configuration_file()
-            )
-
+            # Run pre-hooks
             hooks = self.cli_configuration.hooks
             logger.debug("Running pre-configure apply hook")
             hooks.pre_configure_apply(ctx)
 
-            # Commit the changes made to the config repo
-            ConfigurationGitRepository(cli_context).commit_changes(message)
-
+            # Apply changes
             logger.debug("Applying configuration")
-            self.__generate_configuration_files(configuration, cli_context)
+            ConfigurationManager(cli_context, self.cli_configuration).apply(message)
 
+            # Run post-hooks
             logger.debug("Running post-configure apply hook")
             hooks.post_configure_apply(ctx)
-
-            # Put the generated config repo under version control
-            GeneratedConfigurationGitRepository(cli_context).init()
 
             logger.info("Finished applying configuration")
 
@@ -147,11 +130,11 @@ class ConfigureCli:
         def get(ctx, setting):
             cli_context: CliContext = ctx.obj
 
+            # Validate environment
             self.__pre_configure_get_and_set_validation(cli_context)
 
-            configuration = ConfigurationManager(
-                cli_context.get_app_configuration_file()
-            )
+            # Get settings value and print
+            configuration = ConfigurationManager(cli_context, self.cli_configuration)
             print(configuration.get(setting))
 
         @configure.command(help="Saves a setting to the configuration.")
@@ -161,14 +144,14 @@ class ConfigureCli:
         def set(ctx, setting, value):
             cli_context: CliContext = ctx.obj
 
+            # Validate environment
             self.__pre_configure_get_and_set_validation(cli_context)
 
-            configuration = ConfigurationManager(
-                cli_context.get_app_configuration_file()
-            )
+            # Set settings value
+            configuration = ConfigurationManager(cli_context, self.cli_configuration)
             configuration.set(setting, value)
-            configuration.save()
 
+        # Expose the commands
         self.commands = {"configure": configure}
 
     # ------------------------------------------------------------------------------
@@ -197,216 +180,6 @@ class ConfigureCli:
 
         if has_errors:
             error_and_exit("Missing mandatory environment variables.")
-
-    def __initialise_configuration_directory(self, cli_context: CliContext):
-
-        config_repo: ConfigurationGitRepository = ConfigurationGitRepository(
-            cli_context
-        )
-        app_version: str = cli_context.app_version
-
-        # Initialise the empty repo
-        config_repo.init()
-
-        # Create a new branch for this current application version
-        config_repo.checkout_new_branch(app_version)
-
-        # Seed the configuration directory
-        self.__seed_configuration_dir(cli_context)
-
-        # Commit the changes, and tag as $VERSION
-        config_repo.commit_changes(f"Default configuration at version [{app_version}]")
-        config_repo.tag_current_commit(app_version)
-
-    def __seed_configuration_dir(self, cli_context: CliContext):
-        """Seed the raw configuration into the configuration directory
-
-        Args:
-            cli_context (CliContext): the current cli context
-        """
-        print_header("Seeding configuration directory ...")
-
-        logger.info("Copying app configuration file ...")
-        seed_app_configuration_file = self.cli_configuration.seed_app_configuration_file
-        if not seed_app_configuration_file.is_file():
-            error_and_exit(
-                f"Seed file [{seed_app_configuration_file}] is not valid. Release is corrupt."
-            )
-
-        target_app_configuration_file = cli_context.get_app_configuration_file()
-        logger.debug(
-            "Copying app configuration file to [%s] ...", target_app_configuration_file
-        )
-        target_app_configuration_file.parent.mkdir(parents=True, exist_ok=True)
-        shutil.copy2(seed_app_configuration_file, target_app_configuration_file)
-
-        logger.info("Copying templates ...")
-        templates_dir = cli_context.get_templates_dir()
-        templates_dir.mkdir(parents=True, exist_ok=True)
-        seed_templates_dir = self.cli_configuration.seed_templates_dir
-        if not seed_templates_dir.is_dir():
-            error_and_exit(
-                f"Seed templates directory [{seed_templates_dir}] is not valid. Release is corrupt."
-            )
-
-        for source_file in seed_templates_dir.glob("**/*"):
-            logger.info(source_file)
-            relative_file = source_file.relative_to(seed_templates_dir)
-            target_file = templates_dir.joinpath(relative_file)
-
-            if source_file.is_dir():
-                logger.debug("Creating directory [%s] ...", target_file)
-                target_file.mkdir(parents=True, exist_ok=True)
-            else:
-                logger.debug("Copying seed file to [%s] ...", target_file)
-                shutil.copy2(source_file, target_file)
-
-    def __generate_configuration_files(
-        self, configuration: Configuration, cli_context: CliContext
-    ):
-        """Generate the generated configuration files
-
-        Args:
-            configuration (Configuration): the current cli configuration
-            cli_context (CliContext): the current cli context
-        """
-        print_header(f"Generating configuration files")
-        generated_configuration_dir = cli_context.get_generated_configuration_dir()
-
-        # If the generated configuration directory is not empty, back it up and delete
-        if os.path.exists(generated_configuration_dir) and os.listdir(
-            generated_configuration_dir
-        ):
-            self._backup_and_remove_directory(generated_configuration_dir)
-
-        generated_configuration_dir.mkdir(parents=True, exist_ok=True)
-
-        configuration_record_file = get_generated_configuration_metadata_file(
-            cli_context
-        )
-        if os.path.exists(configuration_record_file):
-            logger.info("Clearing successful configuration record ...")
-            os.remove(configuration_record_file)
-            logger.debug(
-                f"Configuration record removed from [{configuration_record_file}]"
-            )
-
-        for template_file in cli_context.get_templates_dir().glob("**/*"):
-            relative_file = template_file.relative_to(cli_context.get_templates_dir())
-            target_file = generated_configuration_dir.joinpath(relative_file)
-
-            if template_file.is_dir():
-                logger.debug("Creating directory [%s] ...", target_file)
-                target_file.mkdir(parents=True, exist_ok=True)
-                continue
-
-            if template_file.suffix == ".j2":
-                # parse jinja2 templates against configuration
-                target_file = target_file.with_suffix("")
-                logger.info("Generating configuration file [%s] ...", target_file)
-                self.__generate_from_template(
-                    template_file, target_file, configuration.get_as_dict()
-                )
-            else:
-                logger.info("Copying configuration file to [%s] ...", target_file)
-                shutil.copy2(template_file, target_file)
-
-        files_to_decrypt = self.cli_configuration.decrypt_generated_files
-        if len(files_to_decrypt) > 0:
-            self.__decrypt_generated_files(
-                cli_context.get_key_file(),
-                cli_context.get_generated_configuration_dir(),
-                files_to_decrypt,
-            )
-
-        self.__copy_settings_file_to_generated_dir(cli_context)
-
-        logger.info("Saving successful configuration record ...")
-        record = {
-            "generated_at": datetime.utcnow().replace(tzinfo=timezone.utc).isoformat(),
-            "generated_from_commit": ConfigurationGitRepository(
-                cli_context
-            ).get_current_commit_hash(),
-        }
-        configuration_record_file.write_text(
-            json.dumps(record, indent=2, sort_keys=True)
-        )
-        logger.debug("Configuration record written to [%s]", configuration_record_file)
-
-    def _backup_and_remove_directory(self, source_dir: Path):
-        """Backs up a directory to a tar gzipped file with the current datetimestamp,
-        and deletes the existing directory
-
-        Args:
-            source_dir (Path): Path to the directory to backup and delete
-        """
-
-        # The datetime is accurate to seconds (microseconds was overkill), and we remove
-        # colon (:) because `tar tvf` doesn't like filenames with colons
-        current_datetime = (
-            datetime.now().replace(microsecond=0).isoformat().replace(":", "")
-        )
-        basename = os.path.basename(source_dir)
-        output_filename = os.path.join(
-            os.path.dirname(source_dir), f"{basename}.{current_datetime}.tgz"
-        )
-
-        # Create the backup
-        logger.info(
-            f"Backing up current generated configuration directory [{source_dir}] to [{output_filename}]"
-        )
-        with tarfile.open(output_filename, "w:gz") as tar:
-            tar.add(source_dir, arcname=os.path.basename(source_dir))
-
-        # Ensure the backup has been successfully created before deleting the existing generated configuration directory
-        if not os.path.exists(output_filename):
-            error_and_exit(
-                f"Current generated configuration directory backup failed. Could not write out file [{output_filename}]."
-            )
-
-        # Remove the existing directory
-        shutil.rmtree(source_dir, ignore_errors=True)
-        logger.info(
-            f"Deleted previous generated configuration directory [{source_dir}]"
-        )
-
-    def __decrypt_generated_files(
-        self, key_file: Path, generated_config_dir: Path, files: Iterable[str]
-    ):
-        """
-        Decrypts the specified files in the generated configuration
-        directory. The current encrypted version will be overwritten by the
-        decrypted version.
-
-        Args:
-            key_file (Path): Key file to use when decrypting.
-            generated_config_dir (Path): Path to the generated configuration directory.
-            files (Iterable[str]): Relative path to the files to decrypt. Resolved against the generated configuration directory.
-        """
-        for relative_file in files:
-            # decrypt and overwrite the file
-            target_file = generated_config_dir.joinpath(relative_file)
-            logger.debug("Decrypting [%s] ...", target_file)
-            decrypt_values_in_file(target_file, target_file, key_file)
-
-    def __copy_settings_file_to_generated_dir(self, cli_context: CliContext):
-        """Copies the current settings file to the generated directory as a record of what configuration
-        was used to generate those files.
-
-        Args:
-            cli_context (CliContext): The context of the currently-running cli
-        """
-        logger.debug(
-            "Copying applied settings file to generated configuration directory"
-        )
-        applied_configuration_file = cli_context.get_generated_configuration_dir().joinpath(
-            cli_context.get_app_configuration_file().name
-        )
-        shutil.copy2(
-            cli_context.get_app_configuration_file(), applied_configuration_file
-        )
-
-        logger.debug("Applied settings written to [%s]", applied_configuration_file)
 
     def __pre_configure_init_validation(self, cli_context: CliContext):
         """Ensures the system is in a valid state for 'configure init'.
@@ -442,8 +215,8 @@ class ConfigureCli:
             "Checking system configuration is valid before 'configure apply' ..."
         )
 
-        # If the config dir doesn't exist, we cannot apply
-        must_succeed_checks = [confirm_config_dir_exists]
+        # If the config dir doesn't exist, or we're on the master branch, we cannot apply
+        must_succeed_checks = [confirm_config_dir_exists, confirm_not_on_master_branch]
 
         should_succeed_checks = []
 
@@ -484,30 +257,3 @@ class ConfigureCli:
         )
 
         logger.info("System configuration is valid")
-
-    def __generate_from_template(
-        self, template_file: Path, target_file: Path, variables: dict
-    ):
-        """
-        Generate configuration file from the specified template file using
-        the supplied variables.
-
-        Args:
-            template_file (Path): Template used to generate the file.
-            target_file (Path): Location to write the generated file to.
-            variables (dict): Variables used to populate the template.
-        """
-        template = Template(
-            template_file.read_text(),
-            undefined=StrictUndefined,
-            trim_blocks=True,
-            lstrip_blocks=True,
-        )
-        try:
-            output_text = template.render(variables)
-            target_file.write_text(output_text)
-        except Exception as e:
-            error_and_exit(
-                f"Could not generate file from template. The configuration file is likely missing a setting: {e}"
-            )
-
