+++ appcli/commands/migrate_cli.py
@@ -17,17 +17,16 @@ www.brightsparklabs.com
 import click
 
 # local libraries
-from appcli.configuration_manager import ConfigurationManager
-from appcli.functions import execute_validation_functions
-from appcli.logger import logger
-from appcli.models.configuration import Configuration
-from appcli.models.cli_context import CliContext
-from appcli.git_repositories.git_repositories import (
-    ConfigurationGitRepository,
+from appcli.configuration_manager import (
+    ConfigurationManager,
     confirm_config_dir_is_not_dirty,
     confirm_generated_config_dir_is_not_dirty,
     confirm_generated_configuration_is_using_current_configuration,
 )
+from appcli.functions import execute_validation_functions
+from appcli.logger import logger
+from appcli.models.configuration import Configuration
+from appcli.models.cli_context import CliContext
 
 # ------------------------------------------------------------------------------
 # CLASSES
@@ -51,27 +50,13 @@ class MigrateCli:
 
             cli_context: CliContext = ctx.obj
 
+            # Validate environment
             self.__pre_migrate_validation(cli_context)
 
-            config_version: str = self._get_config_version(cli_context)
-            app_version: str = self._get_app_version(cli_context)
+            # Perform migration
+            ConfigurationManager(cli_context, self.cli_configuration).migrate()
 
-            logger.info(
-                f"Migrating configuration at version [{config_version}] to match application version [{app_version}]"
-            )
-
-            # If the configuration version matches the application version, no migration is required.
-            if config_version == app_version:
-                logger.info("Migration not required.")
-                return
-
-            # TODO: Should we have a prompt to confirm that the user definitely wants to migrate from X version to Y version? Include a '-y' or '--yes' option to skip the prompt.
-
-            self._migrate_configuration(cli_context)
-
-            logger.info(
-                f"Migration successfully completed. Migrated configuration from version [{config_version}] to [{app_version}]"
-            )
+            logger.info(f"Migration successfully completed.")
 
         # expose the cli command
         self.commands = {"migrate": migrate}
@@ -98,106 +83,3 @@ class MigrateCli:
         )
 
         logger.info("System configuration is valid for migration.")
-
-    def _get_app_version(self, cli_context: CliContext) -> str:
-        """Get the target application version, which is the version of the application
-        which is currently running in the Docker container.
-
-        Returns:
-            str: version of the application according to the Docker container this script is running in.
-        """
-        return cli_context.app_version
-
-    def _get_config_version(self, cli_context: CliContext) -> str:
-        """Get the current configuration repository's version
-
-        Returns:
-            str: version of the configuration repository
-        """
-        # TODO: return the branch name of the configuration repository
-        config_repo: ConfigurationGitRepository = ConfigurationGitRepository(
-            cli_context
-        )
-        return config_repo.get_current_branch_name()
-
-    def _is_ready_to_migrate(self, cli_context: CliContext) -> bool:
-        """Check if the application is in a state where it is ready to migrate cleanly
-
-        Returns:
-            bool: returns True if the application is ready for migration, otherwise False.
-        """
-        try:
-            confirm_config_dir_is_not_dirty(cli_context)
-        except Exception:
-            logger.error(
-                "Un-applied changes to configuration repository, so migration cannot continue. Please run 'configure apply'."
-            )
-            return False
-
-        return True
-
-    def _migrate_configuration(self, cli_context: CliContext):
-        """Migrates the configuration version to the current application version
-        """
-
-        # TODO: Test and confirm this works as expected.
-
-        config_repo: ConfigurationGitRepository = ConfigurationGitRepository(
-            cli_context
-        )
-        app_version = self._get_app_version(cli_context)
-
-        if config_repo.does_branch_exist(app_version):
-            # If the branch already exists, then this version has previously been installed.
-
-            # TODO: Handle the case where it was previously installed, but we want multiple versions of a single version. e.g. v1-a, v1-b, v1-c
-            logger.warn(
-                f"Version [{app_version}] of this application was previously installed. Rolling back to previous configuration. Manual remediation may be required."
-            )
-
-            # Change to that branch, no further migration steps will be taken.
-            config_repo.checkout_existing_branch(app_version)
-
-            return
-
-        # Read the current configuration variables
-        app_config_file = cli_context.get_app_configuration_file()
-        config_manager: ConfigurationManager = ConfigurationManager(app_config_file)
-        current_variables = config_manager.get_as_dict()
-
-        # Delegate migration to the application callback function
-        new_variables = self.cli_configuration.hooks.migrate_variables(
-            current_variables, self._get_config_version(cli_context)
-        )
-
-        # Change branch to the clean 'master' branch
-        config_repo.checkout_master_branch()
-
-        # TODO: Copy over new version configuration files (re-seed)
-
-        # Create new branch, named after the version being deployed
-        config_repo.checkout_new_branch(app_version)
-
-        # Commit the default v2 configuration
-        config_repo.commit_changes(
-            f"Initialised application at version [{app_version}]"
-        )
-
-        # TODO: Compare new v2 config to the 'clean v2', and make sure all settings are there.
-
-        # Write out 'v2 migrated' variables file
-        config_manager.set_all(new_variables)
-        config_manager.save()
-
-        # Commit the new variables file
-        config_repo.commit_changes(
-            f"Migrated variables file to version [{app_version}]"
-        )
-
-        # >> Now at v2 variables file. Still templates to go.
-
-        # TODO: Diff all non-variables files (i.e. all templates files + w/e else) -> List of changed files, and their changes.
-
-        # TODO: For each changed file, notify user that manual changes will need to be made to that template file. Provide diff.
-
-        # TODO: Once all templates have been upgraded to v2, commit and done with migration!
