+++ appcli/functions.py
@@ -11,7 +11,6 @@ www.brightsparklabs.com
 
 # standard libraries
 import click
-from pathlib import Path
 import re
 import sys
 from typing import Callable, Iterable, List
@@ -19,13 +18,6 @@ from typing import Callable, Iterable, List
 # local libraries
 from appcli.logger import logger
 from appcli.models.cli_context import CliContext
-
-# ------------------------------------------------------------------------------
-# CONSTANTS
-# ------------------------------------------------------------------------------
-
-METADATA_FILE_NAME = "metadata-configure-apply.json"
-""" Name of the file holding metadata from running a configure (relative to the generated configuration directory) """
 
 # ------------------------------------------------------------------------------
 # VARIABLES
@@ -57,11 +49,6 @@ def print_header(title):
                         ============================================================""",
         title.upper(),
     )
-
-
-def get_generated_configuration_metadata_file(cli_context: CliContext) -> Path:
-    generated_configuration_dir = cli_context.get_generated_configuration_dir()
-    return generated_configuration_dir.joinpath(METADATA_FILE_NAME)
 
 
 def extract_valid_environment_variable_names(
