+++ appcli/git_repositories/git_repositories.py
@@ -10,8 +10,6 @@ www.brightsparklabs.com
 """
 
 # standard library
-import json
-import os
 from pathlib import Path
 from typing import Iterable
 
@@ -19,12 +17,17 @@ from typing import Iterable
 import git
 
 # local libraries
-from appcli.functions import get_generated_configuration_metadata_file
 from appcli.logger import logger
 from appcli.models.cli_context import CliContext
 
 # ------------------------------------------------------------------------------
-# PRIVATE CLASSES
+# CONSTANTS
+# ------------------------------------------------------------------------------
+
+BASE_BRANCH_NAME: str = "master"
+
+# ------------------------------------------------------------------------------
+# PUBLIC CLASSES
 # ------------------------------------------------------------------------------
 
 
@@ -32,9 +35,8 @@ class GitRepository:
     """Class which encapsulates different git repo actions for configuration repositories
     """
 
-    def __init__(self, repo_path: str, ignores: Iterable[str] = None):
+    def __init__(self, repo_path: str):
         self.repo_path = repo_path
-        self.ignores = ignores
         self.actor: git.Actor = git.Actor(f"appcli", "root@localhost")
 
     def init(self):
@@ -51,19 +53,13 @@ class GitRepository:
         # git init, and write to the .gitignore file
         repo = git.Repo.init(self.repo_path)
         logger.debug("Initialised repository at [%s]", repo.working_dir)
-        with open(self.repo_path.joinpath(".gitignore"), "w+") as ignore_file:
-            if self.ignores:
-                for ignore in self.ignores:
-                    ignore_file.write(f"{ignore}\n")
-        logger.debug("Created .gitignore with ignores: [%s]", self.ignores)
-        repo.git.add(".gitignore")
 
-        # TODO: Generate encryption key here on the master branch, it should be used across all branches
-
-        # do the initial commit on the repo
-        repo.git.add("*")
-        repo.index.commit("[autocommit] Initialised repository", author=self.actor)
-        logger.debug("Committed repository at [%s]", repo.working_dir)
+    def set_gitignore(self, ignores: Iterable[str]):
+        # Overwrite existing .gitignore file
+        with open(self.repo_path.joinpath(".gitignore"), "w") as ignore_file:
+            for ignore in ignores:
+                ignore_file.write(f"{ignore}\n")
+        logger.debug("Created .gitignore with ignores: [%s]", ignores)
 
     def commit_changes(self, message: str):
         """Commit the existing changes to the git repository
@@ -188,166 +184,11 @@ class GitRepository:
         return repo
 
 
-# ------------------------------------------------------------------------------
-# CLASSES
-# ------------------------------------------------------------------------------
-
-
 class ConfigurationGitRepository(GitRepository):
     def __init__(self, cli_context: CliContext):
-        super().__init__(cli_context.configuration_dir, [".generated*"])
+        super().__init__(cli_context.configuration_dir)
 
 
 class GeneratedConfigurationGitRepository(GitRepository):
     def __init__(self, cli_context: CliContext):
         super().__init__(cli_context.get_generated_configuration_dir())
-
-
-# ------------------------------------------------------------------------------
-# FUNCTIONS
-# ------------------------------------------------------------------------------
-
-
-def confirm_config_dir_exists(cli_context: CliContext):
-    """Confirm that the configuration repository exists.
-    If this fails, it will raise a general Exception with the error message.
-
-    Args:
-        cli_context (CliContext): the current cli context
-
-    Raises:
-        Exception: Raised if the configuration repository does *not* exist.
-    """
-    config_repo: ConfigurationGitRepository = ConfigurationGitRepository(cli_context)
-    if not config_repo.repo_exists():
-        raise Exception(
-            f"Configuration does not exist at [{config_repo.repo_path}]. Please run `configure init`."
-        )
-
-
-def confirm_config_dir_not_exists(cli_context: CliContext):
-    """Confirm that the configuration repository does *not* exist.
-    If this fails, it will raise a general Exception with the error message.
-
-    Args:
-        cli_context (CliContext): the current cli context
-
-    Raises:
-        Exception: Raised if the configuration repository exists.
-    """
-    config_repo: ConfigurationGitRepository = ConfigurationGitRepository(cli_context)
-    if config_repo.repo_exists():
-        raise Exception(f"Configuration already exists at [{config_repo.repo_path}].")
-
-
-def confirm_generated_config_dir_exists(cli_context: CliContext):
-    """Confirm that the generated configuration repository exists.
-    If this fails, it will raise a general Exception with the error message.
-
-    Args:
-        cli_context (CliContext): the current cli context
-
-    Raises:
-        Exception: Raised if the generated configuration repository does not exist.
-    """
-    generated_config_repo: GeneratedConfigurationGitRepository = GeneratedConfigurationGitRepository(
-        cli_context
-    )
-    if not generated_config_repo.repo_exists():
-        raise Exception(
-            f"Generated configuration does not exist at [{generated_config_repo.repo_path}]. Please run `configure apply`."
-        )
-
-
-def confirm_config_dir_is_not_dirty(cli_context: CliContext):
-    """Confirm that the configuration repository has not been modified and not 'applied'.
-    If this fails, it will raise a general Exception with the error message.
-
-    Args:
-        cli_context (CliContext): the current cli context
-
-    Raises:
-        Exception: Raised if the configuration repository has been modified and not 'applied'.
-    """
-    config_repo: ConfigurationGitRepository = ConfigurationGitRepository(cli_context)
-    if config_repo.is_dirty(untracked_files=True):
-        raise Exception(
-            f"Configuration at [{config_repo.repo_path}]] contains changes which have not been applied. Please run `configure apply`."
-        )
-
-
-def confirm_generated_config_dir_is_not_dirty(cli_context: CliContext):
-    """Confirm that the generated configuration repository has not been manually modified and not checked-in.
-    If this fails, it will raise a general Exception with the error message.
-
-    Args:
-        cli_context (CliContext): the current cli context
-
-    Raises:
-        Exception: Raised if the generated configuration repository has been manually modified and not checked in.
-    """
-    generated_config_repo: GeneratedConfigurationGitRepository = GeneratedConfigurationGitRepository(
-        cli_context
-    )
-    if generated_config_repo.is_dirty(untracked_files=False):
-        raise Exception(
-            f"Generated configuration at [{generated_config_repo.repo_path}] has been manually modified."
-        )
-
-
-def confirm_generated_configuration_is_using_current_configuration(
-    cli_context: CliContext,
-):
-    """Confirm that the generated configuration directory was generated from the current state configuration directory.
-    If this fails, it will raise a general Exception with the error message.
-
-    Args:
-        cli_context (CliContext): the current cli context
-
-    Raises:
-        Exception: Raised if metadata file not found, or generated config is out of sync with config.
-    """
-    metadata_file = get_generated_configuration_metadata_file(cli_context)
-    if not os.path.isfile(metadata_file):
-        raise Exception(
-            f"Could not find a metadata file at [{metadata_file}]. Please run `configure apply`"
-        )
-
-    with open(metadata_file, "r") as f:
-        metadata = json.load(f)
-        logger.debug("Metadata from generated configuration: %s", metadata)
-
-    generated_commit_hash = metadata["generated_from_commit"]
-    configuration_commit_hash = ConfigurationGitRepository(
-        cli_context
-    ).get_current_commit_hash()
-    if generated_commit_hash != configuration_commit_hash:
-        logger.debug(
-            "Generated configuration hash [%s] does not match configuration hash [%s]",
-            generated_commit_hash,
-            configuration_commit_hash,
-        )
-        raise Exception(
-            "Generated configuration is out of sync with raw configuration. Please run `configure apply`."
-        )
-
-
-def confirm_config_version_matches_app_version(cli_context: CliContext):
-    """Confirm that the configuration repository version matches the application version.
-    If this fails, it will raise a general Exception with the error message.
-
-    Args:
-        cli_context (CliContext): the current cli context
-
-    Raises:
-        Exception: Raised if the configuration repository version doesn't match the application version.
-    """
-    config_repo: ConfigurationGitRepository = ConfigurationGitRepository(cli_context)
-    config_version: str = config_repo.get_current_branch_name()
-
-    app_version: str = cli_context.app_version
-
-    if config_version != app_version:
-        raise Exception(
-            f"Configuration at [{config_repo.repo_path}] is using version [{config_version}] which is incompatible with current application version [{app_version}]. Migrate to this application version using 'migrate'."
-        )
